server/requirements.txt
requests
Django
gunicorn==20.0.4
Pillow==8.0.1
python-dotenv

